Separate out strategies, fix some improper usages
