+++ pandas/tests/indexing/multiindex/conftest.py
@@ -22,6 +22,7 @@ def multiindex_year_month_day_dataframe_random_data():
     """DataFrame with 3 level MultiIndex (year, month, day) covering
     first 100 business days from 2000-01-01 with random data"""
     tm.N = 100
+    tm.strategies.N = 100
     tdf = tm.makeTimeDataFrame()
     ymd = tdf.groupby([lambda x: x.year, lambda x: x.month,
                        lambda x: x.day]).sum()
