+++ pandas/tests/series/test_combine_concat.py
@@ -90,7 +90,7 @@ class TestSeriesCombine(object):
 
         # mixed types
         index = tm.makeStringIndex(20)
-        floats = Series(tm.randn(20), index=index)
+        floats = Series(np.random.randn(20), index=index)
         strings = Series(tm.makeStringIndex(10), index=index[::2])
 
         combined = strings.combine_first(floats)
